new authController and new AuthRoutes
