+++ backend/src/controllers/authController.ts
@@ -3,10 +3,6 @@ import User from "../models/user"
 
 import bcrypt from "bcryptjs"
 import { DuplicateEmailError } from "../utils/DuplicateEmailError"
-import Tour from "../models/tour"
-import Profile from "../models/profile"
-import { ProfileProps, UserProps } from "../types/Profile"
-import mongoose from "mongoose"
 
 require("dotenv").config()
 
@@ -97,170 +93,5 @@ export const logout = async (req: Request, res: Response) => {
 		const err = error as Error
 		console.error("Error logging out user:", err.message)
 		res.status(500).json({ error: "Internal server error" })
-	}
-}
-
-export const getAllUsers = async (_req: Request, res: Response) => {
-	try {
-		const users = await User.find()
-		res.status(200).json({ users })
-	} catch (error) {
-		const err = error as Error
-		res.status(500).json({
-			error: err.message,
-		})
-	}
-}
-
-export const createUserProfile = async (req: Request, res: Response) => {
-	try {
-		const { userId, firstName, lastName, imageUri } = req.body
-
-		const user = await User.findById(userId)
-		if (!user) {
-			return res.status(404).json({ error: "User not found" })
-		}
-
-		const profile = new Profile({
-			user: userId,
-			firstName,
-			lastName,
-			imageUri,
-		})
-
-		await profile.save()
-
-		res.status(201).json({ message: "Profile created successfully" })
-	} catch (error) {
-		const err = error as Error
-		res.status(400).json({ error: err.message })
-	}
-	return undefined
-}
-
-export const getUserProfileById = async (req: Request, res: Response) => {
-	try {
-		const userId = req.params.userId
-
-		const user = await User.findById(userId)
-		if (!user) {
-			return res.status(404).json({
-				error: "User not found",
-			})
-		}
-
-		const profile = await Profile.findOne({ user: userId })
-		if (!profile) {
-			return res.status(404).json({ error: "Profile not found" })
-		}
-
-		res.status(200).json({ profile })
-	} catch (error) {
-		const err = error as Error
-		res.status(500).json({ error: err.message })
-	}
-	return undefined
-}
-
-export const getUserProfile = async (req: Request, res: Response) => {
-	try {
-		const userId = req.params.id
-
-		// Find the user by ID
-		const user = await User.findById(userId)
-		if (!user) {
-			return res.status(404).json({ error: "User not found" })
-		}
-
-		// Find the associated profile
-		const profile = await Profile.findOne({ user: userId })
-		if (!profile) {
-			return res.status(404).json({ error: "Profile not found" })
-		}
-
-		// Construct the user profile response
-		const userProfile = {
-			_id: user._id,
-			email: user.email,
-			firstName: profile.firstName || "",
-			lastName: profile.lastName || "",
-			imageUri: profile.imageUri || "",
-		}
-
-		// Return the combined user and profile data
-		res.json({ user: userProfile })
-	} catch (error) {
-		console.error("Error fetching user profile:", error) // Log error for debugging
-		res.status(500).json({
-			error: "An unexpected error occurred while fetching the user profile",
-		})
-	}
-}
-
-export const updateUserProfile = async (
-	req: Request,
-	res: Response,
-	next: NextFunction
-) => {
-	try {
-		const { userId } = req.params
-		const { firstName, lastName, imageUri } = req.body
-
-		// Validate input
-		if (!firstName || !lastName) {
-			return res
-				.status(400)
-				.json({ error: "First Name and Last Name are required" })
-		}
-
-		// Find user and update profile
-		const user = (await User.findById(userId)) as UserProps
-		if (!user) {
-			return res.status(404).json({ error: "User not found" })
-		}
-
-		let profile = await Profile.findOne({ user: userId })
-		if (!profile) {
-			profile = new Profile({ user: userId, firstName, lastName, imageUri })
-		} else {
-			profile.firstName = firstName
-			profile.lastName = lastName
-			profile.imageUri = imageUri
-		}
-
-		await profile.save()
-
-		res.status(200).json({ message: "Profile updated successfully", profile })
-	} catch (error) {
-		const err = error as Error
-		res.status(400).json({ error: err.message })
-		next(error)
-	}
-}
-
-export const deleteUser = async (req: Request, res: Response) => {
-	try {
-		const userId = req.params.userId
-
-		const user = await User.findByIdAndDelete(userId).populate("profile")
-
-		if (!user) {
-			return res.status(404).json({ error: "User not found" })
-		}
-
-		// Check if a profile is associated with the user
-		if (user.profile) {
-			// await Profile.findByIdAndDelete(user.) // Delete the profile
-		}
-
-		// Delete related tours (assuming a Tour model exists)
-		await Tour.deleteMany({ creator: userId })
-
-		res
-			.status(204)
-			.json({ message: "User and associated data deleted successfully" })
-	} catch (error) {
-		const err = error as Error
-		res.status(500).json({ error: err.message })
 	}
 }
